requirements.txt
PyYAML
pandas
XlsxWriter
dataclasses; python_version<"3.7"
